+++ MatrixKit/Controllers/MXKRoomViewController.m
@@ -53,7 +53,9 @@
 #import "MXKSlashCommands.h"
 #import "MXKSwiftHeader.h"
 
-@interface MXKRoomViewController ()
+#import "MXKPreviewViewController.h"
+
+@interface MXKRoomViewController () <MXKPreviewViewControllerDelegate>
 {
     /**
      YES if scrolling to bottom is in progress
@@ -213,6 +215,9 @@
     
     // Keep visible the status bar by default.
     isStatusBarHidden = NO;
+    
+    // By default actions button is shown in document preview
+    _allowActionsInDocumentPreview = YES;
 }
 
 - (void)viewDidLoad
@@ -283,6 +288,8 @@
             self.bubbleTableViewDisplayInTransition = NO;
         }
     }];
+    
+    self.navigationController.delegate = self;
 }
 
 - (BOOL)prefersStatusBarHidden
@@ -2307,8 +2314,10 @@
             UITableViewCell *cellTmp;
             if (!cell)
             {
+                NSString *reuseIdentifier = [self cellReuseIdentifierForCellData:[roomDataSource cellDataAtIndex:rowIndex]];
                 // Create temporarily the cell (this cell will released at the end, to be reusable)
-                cellTmp = [roomDataSource tableView:_bubblesTableView cellForRowAtIndexPath:[NSIndexPath indexPathForRow:rowIndex inSection:0]];
+                cellTmp = [_bubblesTableView dequeueReusableCellWithIdentifier:reuseIdentifier
+                                                                  forIndexPath:[NSIndexPath indexPathForRow:rowIndex inSection:0]];
                 cell = cellTmp;
             }
             
@@ -3708,18 +3717,29 @@
 
                     void(^viewAttachment)(void) = ^() {
 
-                        self->documentInteractionController = [UIDocumentInteractionController interactionControllerWithURL:fileURL];
-                        [self->documentInteractionController setDelegate:self];
-                        self->currentSharedAttachment = selectedAttachment;
-
-                        if (![self->documentInteractionController presentPreviewAnimated:YES])
+                        if (self.allowActionsInDocumentPreview)
                         {
-                            if (![self->documentInteractionController presentOptionsMenuFromRect:self.view.frame inView:self.view animated:YES])
+                            // We could get rid of this part of code and use only a MXKPreviewViewController
+                            // Nevertheless, MXKRoomViewController is compliant to UIDocumentInteractionControllerDelegate
+                            // and remove all this code could have effect on some custom implementations.
+                            self->documentInteractionController = [UIDocumentInteractionController interactionControllerWithURL:fileURL];
+                            [self->documentInteractionController setDelegate:self];
+                            self->currentSharedAttachment = selectedAttachment;
+
+                            if (![self->documentInteractionController presentPreviewAnimated:YES])
                             {
-                                self->documentInteractionController = nil;
-                                [selectedAttachment onShareEnded];
-                                self->currentSharedAttachment = nil;
+                                if (![self->documentInteractionController presentOptionsMenuFromRect:self.view.frame inView:self.view animated:YES])
+                                {
+                                    self->documentInteractionController = nil;
+                                    [selectedAttachment onShareEnded];
+                                    self->currentSharedAttachment = nil;
+                                }
                             }
+                        }
+                        else
+                        {
+                            self->currentSharedAttachment = selectedAttachment;
+                            [MXKPreviewViewController presentFrom:self fileUrl:fileURL allowActions:self.allowActionsInDocumentPreview delegate:self];
                         }
                     };
 
@@ -3833,6 +3853,17 @@
     if (shouldScrollToBottom)
     {
         [self scrollBubblesTableViewToBottomAnimated:YES];
+    }
+}
+
+#pragma mark - MXKPreviewViewControllerDelegate
+
+- (void)previewViewControllerDidEndPreview:(MXKPreviewViewController *)controller
+{
+    if (currentSharedAttachment)
+    {
+        [currentSharedAttachment onShareEnded];
+        currentSharedAttachment = nil;
     }
 }